Re-implementing moveNext and movePrevious with ko
It now updates the items in the container when it has finished sliding
(drops one of one end, adds a new to the other)

This also required dealing with child items in the container firing
transition completed events. The solution works on the assumption that
the child items have their own bindings and viewmodel so I can check
the ko.dataFor of the event target. (This felt nicer than the view
model needing to know which element it was getting bound to.)

Also, updated slidingStatus to be a computed property based on the
direction that the container is moving.

TODO:
- fading in/out of the items that are getting added/removed
- setting the current item