Accelerate the animations
So you can hold down the arrow keys and start to whizz through the items

Instead of specifying the animation duration in CSS, it's bound to a view
model computed property, which is determined from the slidingStatus and
a time value.

The time value decreses after each move but resets to the initial time
if there are no moves within a timeout.

It works ok for holding down the arrows to speed up moving through the
items, but still feels a bit janky when hitting the arrows multiple times
to move through quickly. It'll do for now.

Also - the tests are a right mess now...I need to something about that
pretty soon.